Added marker generation from list of objects
Added auto-generation of markers on map from a list of objects with a location in them. This data be gathered later from a MongoDB query. For now, there is on click behavior and on double click behavior with little functionality, but this can be tweaked to get more information later.