Added onClick functionality to Map
Added onClick functionality to the Map component. When a user clicks a marker, a modal will appear to give them information based on the data from the marker. I also allowed the other components to be accessible if properly routed to. Note: The routing needs to be fixed, but I don't want to do this until the database work is done.